+++ src/wmts/wmts_plugin.cpp
@@ -91,7 +91,7 @@ QcWmtsPluginLayer::create_tile_spec(int level, int x, int y) const
 
 /**************************************************************************************************/
 
-QcWmtsPlugin::QcWmtsPlugin(const QString & name, const QString & title, size_t number_of_levels, size_t tile_size)
+QcWmtsPlugin::QcWmtsPlugin(const QString & name, const QString & title, int number_of_levels, int tile_size)
   : m_name(name),
     m_title(title),
     m_tile_matrix_set(name, number_of_levels, tile_size),
